[AIP-134] Document AIP-134 rules.
