+++ docs/rules/core/0134-request-message-name.md
@@ -1,18 +1,18 @@
 ---
 rule:
-  aip: 133
-  name: [core, '0133', request-message-name]
-  summary: Create methods must have standardized request message names.
+  aip: 134
+  name: [core, '0134', request-message-name]
+  summary: Update methods must have standardized request message names.
 ---
 
-# Create methods: Request message
+# Update methods: Request message
 
-This rule enforces that all `Create` RPCs have a request message name of
-`Create*Request`, as mandated in [AIP-133][].
+This rule enforces that all `Update` RPCs have a request message name of
+`Update*Request`, as mandated in [AIP-134][].
 
 ## Details
 
-This rule looks at any message matching beginning with `Create`, and complains
+This rule looks at any message matching beginning with `Update`, and complains
 if the name of the corresponding input message does not match the name of the
 RPC with the suffix `Request` appended.
 
@@ -22,9 +22,9 @@ RPC with the suffix `Request` appended.
 
 ```proto
 // Incorrect.
-rpc CreateBook(Book) returns (Book) {  // Should be `CreateBookRequest`.
+rpc UpdateBook(Book) returns (Book) {  // Should be `UpdateBookRequest`.
   option (google.api.http) = {
-    post: "/v1/{parent=publishers/*}/books"
+    patch: "/v1/{name=publishers/*/books/*}"
     body: "*"
   };
 }
@@ -34,9 +34,9 @@ rpc CreateBook(Book) returns (Book) {  // Should be `CreateBookRequest`.
 
 ```proto
 // Correct.
-rpc CreateBook(CreateBookRequest) returns (Book) {
+rpc UpdateBook(UpdateBookRequest) returns (Book) {
   option (google.api.http) = {
-    post: "/v1/{parent=publishers/*}/books"
+    patch: "/v1/{book.name=publishers/*/books/*}"
     body: "book"
   };
 }
@@ -48,11 +48,11 @@ If you need to violate this rule, use a leading comment above the method.
 Remember to also include an [aip.dev/not-precedent][] comment explaining why.
 
 ```proto
-// (-- api-linter: core::0133::request-message-name=disabled
+// (-- api-linter: core::0134::request-message-name=disabled
 //     aip.dev/not-precedent: We need to do this because reasons. --)
-rpc CreateBook(Book) returns (Book) {
+rpc UpdateBook(Book) returns (Book) {  // Should be `UpdateBookRequest`.
   option (google.api.http) = {
-    post: "/v1/{parent=publishers/*}/books"
+    patch: "/v1/{name=publishers/*/books/*}"
     body: "*"
   };
 }
@@ -61,5 +61,5 @@ rpc CreateBook(Book) returns (Book) {
 If you need to violate this rule for an entire file, place the comment at the
 top of the file.
 
-[aip-133]: https://aip.dev/133
+[aip-134]: https://aip.dev/134
 [aip.dev/not-precedent]: https://aip.dev/not-precedent
