+++ rules/aip0134/request_resource_field_test.go
@@ -12,43 +12,39 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package aip0133
+package aip0134
 
 import (
 	"testing"
 
 	"github.com/googleapis/api-linter/rules/internal/testutils"
 	"github.com/jhump/protoreflect/desc"
-	"github.com/jhump/protoreflect/desc/builder"
 )
 
 func TestResourceField(t *testing.T) {
 	// Set up the testing permutations.
 	tests := []struct {
-		testName     string
-		messageName  string
-		messageField *field
-		descriptor   func(*desc.MessageDescriptor) desc.Descriptor
-		problems     testutils.Problems
+		testName          string
+		MessageName       string
+		ResourceName      string
+		ResourceFieldName string
+		descriptor        func(*desc.MessageDescriptor) desc.Descriptor
+		problems          testutils.Problems
 	}{
+		{"Valid", "UpdateBookRequest", "Book", "book", nil, testutils.Problems{}},
+		{"ValidTwoWords", "UpdateBigBookRequest", "BigBook", "big_book", nil, testutils.Problems{}},
 		{
-			"Valid",
-			"CreateBookRequest",
-			&field{"book", builder.FieldTypeMessage(builder.NewMessage("Book"))},
+			"InvalidMismatch",
+			"UpdateBookRequest",
+			"Foo",
+			"foo",
 			nil,
-			testutils.Problems{},
-		},
+			testutils.Problems{{Message: "has no \"Book\""}}},
 		{
-			"MissingField",
-			"CreateBookRequest",
-			nil,
-			nil,
-			testutils.Problems{{Message: "Message \"CreateBookRequest\" has no \"Book\" type field"}},
-		},
-		{
-			"WrongName",
-			"CreateBookRequest",
-			&field{"payload", builder.FieldTypeMessage(builder.NewMessage("Book"))},
+			"InvalidFieldName",
+			"UpdateBookRequest",
+			"Book",
+			"big_book",
 			func(m *desc.MessageDescriptor) desc.Descriptor {
 				return m.GetFields()[0]
 			},
@@ -59,27 +55,21 @@ func TestResourceField(t *testing.T) {
 	// Run each test individually.
 	for _, test := range tests {
 		t.Run(test.testName, func(t *testing.T) {
-			// Create an appropriate message descriptor.
-			messageBuilder := builder.NewMessage(test.messageName)
-
-			if test.messageField != nil {
-				messageBuilder.AddField(
-					builder.NewField(test.messageField.fieldName, test.messageField.fieldType),
-				)
-			}
-
-			message, err := messageBuilder.Build()
-			if err != nil {
-				t.Fatalf("Could not build %s message.", test.messageName)
-			}
-
-			var d desc.Descriptor = message
+			file := testutils.ParseProto3Tmpl(t, `
+				import "google/protobuf/field_mask.proto";
+				message {{.MessageName}} {
+					{{.ResourceName}} {{.ResourceFieldName}} = 1;
+					google.protobuf.FieldMask update_mask = 2;
+				}
+				message {{.ResourceName}} {}
+			`, test)
+			var d desc.Descriptor = file.GetMessageTypes()[0]
 			if test.descriptor != nil {
-				d = test.descriptor(message)
+				d = test.descriptor(d.(*desc.MessageDescriptor))
 			}
 
 			// Run the lint rule, and establish that it returns the correct problems.
-			problems := resourceField.Lint(message.GetFile())
+			problems := resourceField.Lint(file)
 			if diff := test.problems.SetDescriptor(d).Diff(problems); diff != "" {
 				t.Errorf(diff)
 			}
